refactor(trajecten): extract useLoginToChooser composable
TrajectMenu and MobileTrajectSheet had a verbatim copy of the "log in, then
return to the traject chooser" redirect. Move it to a shared
useLoginToChooser composable so the two can't diverge silently.
