+++ frontend/src/components/TrajectMenu.vue
@@ -4,6 +4,7 @@ import { useRoute, useRouter } from 'vue-router';
 import { useTrajects } from '../composables/useTrajects.js';
 import { useDocumentsSheet } from '../composables/useDocumentsSheet.js';
 import { useAuth } from '../composables/useAuth.js';
+import { useLoginToChooser } from '../composables/useLoginToChooser.js';
 import TrajectMembersDialog from './TrajectMembersDialog.vue';
 import TrajectInfoDialog from './TrajectInfoDialog.vue';
 import TrajectCreateForm from './TrajectCreateForm.vue';
@@ -28,26 +29,14 @@ const {
 } = useTrajects();
 // Auth gates the menu: logged-in users get the traject switcher; everyone
 // else gets a popover explaining that trajecten unlock after login.
-const { authenticated, login } = useAuth();
+const { authenticated } = useAuth();
 const route = useRoute();
 const router = useRouter();
 const documentsSheet = useDocumentsSheet();
 
 // Not logged in: log in, then land on the trajectchooser carrying the current
-// section + law/article, so picking a traject returns the user to where they
-// were — now traject-scoped.
-function loginToChooser() {
-  const inLibrary = route.name === 'library' || route.name === 'library-traject';
-  const target = router.resolve({
-    name: 'trajecten',
-    query: {
-      sectie: inLibrary ? 'library' : 'editor',
-      law: route.params.lawId || undefined,
-      article: route.params.articleNumber || undefined,
-    },
-  });
-  login(target.fullPath);
-}
+// section + law/article (shared composable, see also MobileTrajectSheet).
+const loginToChooser = useLoginToChooser();
 
 /**
  * Navigate to a traject — push the user into the traject-scoped view of
